fix missing values in apps.py
