+++ .circleci/vars/branch_overrides.py
@@ -28,5 +28,9 @@ branch_overrides = {
     "4311-stream-block-error": {
         "LOAD_DATA": "fixtures",
         "V3_WIP": True,
+    },
+    "4356-queue": {
+        "CI_COA_PUBLISHER_V2_URL_PR": "https://oar72z1wgf.execute-api.us-east-1.amazonaws.com/4356-queue/publish-request",
+        "COA_PUBLISHER_V2_API_KEY_PR": "changeme",
     }
 }
